src/streamlit_app/requirements.txt
# Streamlit application dependencies
streamlit>=1.28.0
boto3>=1.34.0
botocore>=1.34.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.5.0
python-multipart>=0.0.6
websockets>=12.0
websocket-client>=1.6.0
pyjwt>=2.8.0
streamlit-autorefresh>=1.0.0
pillow>=10.1.0
numpy>=1.24.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.17.0
altair>=5.2.0
markdown>=3.5.0
watchdog>=3.0.0
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.21.0